requirements.txt
fastapi
uvicorn
jinja2
python-multipart
pymongo
requests
cloudinary
authlib
python-dotenv
